Update
Changed image system to store in images folder.
Changed so that framework supports making larger images out of smaller images.
TODO: Combine small images into one big one.

+++ Point.cs
@@ -10,7 +10,7 @@ namespace MapGenerator
         }
         public long x;
         public long y;
-        public int value = 0;
+        public double value = 0;
         private int r;
         public int R
         {
